+++ app/Jobs/Issues/SendEmailToFollowersAboutReminder.php
@@ -43,21 +43,52 @@ class SendEmailToFollowersAboutReminder implements ShouldQueue
         if (!is_null($this->issue->timeClosed)) {
             return;
         }
-        if ($this->typeOfReminder == 'paused') {
-            if (!$this->issue->paused) {
+
+        switch ($this->typeOfReminder) {
+            case 'paused':
+                if (!$this->issue->paused) {
+                    return null;
+                }
+                $delayDays=setting('days_reminder_paused_issue');
+                break;
+            case 'waitingForInternal':
+                if (!$this->issue->waitingForInternal) {
+                    return null;
+                }
+                $delayDays=setting('days_reminder_waiting_for_internal');
+                break;
+            case 'waitingForCustomer':
+                if (!$this->issue->waitingForCustomer) {
+                    return null;
+                }
+                $delayDays=setting('days_reminder_waiting_for_customer');
+                break;
+                
+            default:
+                $delayDays=setting('days_reminder_issue');
+                break;
+        }
+        // if ($this->typeOfReminder == 'paused') {
+        //     if (!$this->issue->paused) {
+        //         return null;
+        //     }
+            // Has there been comments since the job was created, don't send mail, just add new reminder
+            if ($this->issue->latestComment->updated_at->addDays($delayDays) > now()) {
+                $delay = nextWorkingHour(now()->addDays($delayDays));
+                CreateNewReminder::dispatch($this->issue, $this->typeOfReminder)->delay($delay);
                 return null;
             }
         }
-        if ($this->typeOfReminder == 'waitingForInternal') {
-            if (!$this->issue->waitingForInternal) {
-                return null;
-            }
-        }
-        if ($this->typeOfReminder == 'waitingForCustomer') {
-            if (!$this->issue->waitingForCustomer) {
-                return null;
-            }
-        }
+        // if ($this->typeOfReminder == 'waitingForInternal') {
+        //     if (!$this->issue->waitingForInternal) {
+        //         return null;
+        //     }
+        // }
+        // if ($this->typeOfReminder == 'waitingForCustomer') {
+        //     if (!$this->issue->waitingForCustomer) {
+        //         return null;
+        //     }
+        // }
         Mail::to($this->email)->send(new MailToFollowersAboutReminder($this->issue, $this->typeOfReminder));
     }
 }